+++ src/streamdiffusion/preprocessing/processors/__init__.py
