+++ src/streamdiffusion/preprocessing/processors/base.py
