+++ src/streamdiffusion/preprocessing/processors/canny.py
