+++ src/streamdiffusion/preprocessing/processors/depth.py
