+++ src/streamdiffusion/preprocessing/processors/external.py
