+++ src/streamdiffusion/preprocessing/processors/hed.py
