+++ src/streamdiffusion/preprocessing/processors/ipadapter_embedding.py
