+++ src/streamdiffusion/preprocessing/processors/mediapipe_segmentation.py
