+++ src/streamdiffusion/preprocessing/processors/openpose.py
