+++ src/streamdiffusion/preprocessing/processors/passthrough.py
