+++ src/streamdiffusion/preprocessing/processors/soft_edge.py
